+++ ACP3/Modules/ACP3/System/Resources/Assets/js/admin/index.configuration.js
@@ -1,4 +1,12 @@
 jQuery(document).ready(function ($) {
+    $(':radio[name="cookie_consent_is_enabled"]').on('click change', function () {
+        $('#cookie-consent-container').toggle(this.value == 1);
+    }).filter(':checked').triggerHandler('click');
+
+    $(':radio[name="maintenance_mode"]').on('click change', function () {
+        $('#maintenance-message-container').toggle(this.value == 1);
+    }).filter(':checked').triggerHandler('click');
+
     $(':radio[name="mailer_smtp_auth"]').on('click change', function () {
         $('#mailer-smtp-2').toggle(this.value == 1);
     }).filter(':checked').triggerHandler('click');